Add implicit double/float to FpxxHost conversion methods
